+++ packages/core-js/actual/array-buffer/index.js
@@ -1,3 +1,6 @@
 var parent = require('../../stable/array-buffer');
+require('../../modules/esnext.array-buffer.detached');
+require('../../modules/esnext.array-buffer.transfer');
+require('../../modules/esnext.array-buffer.transfer-to-fixed-length');
 
 module.exports = parent;