+++ packages/core-js/full/array-buffer/transfer-to-fixed-length.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/array-buffer/transfer-to-fixed-length');
+
+module.exports = parent;